+++ packages/core/src/core/effect.ts
@@ -179,6 +179,7 @@ class EffectImpl extends ReactiveNode<void> implements EffectObject, DependencyT
     if (!force && deps.size > 0 && !this._isDirty()) return;
 
     this._checkInfiniteLoops();
+    debug.trackUpdate(this.id, debug.getDebugName(this));
 
     this.flags = flags | EFFECT_STATE_FLAGS.EXECUTING;
     this._execCleanup();